disable on master until prow is enabled for now (#615)
Automatic merge from submit-queue.

disable on master until prow is enabled for now

The bot is sending spam emails and is not useful because Jenkins is broken on mixerclient
and proxy. We need to enable it again, when we are ready. For now dev agreed to manually update SHA.
**Release note**:

```release-note
none
```