Change pkg_check to return 0 for package violation.
So that the codecov job will not fail. Otherwise, PR authors can be
easily confused since job failure can be caused by others and the error
is not actionable by the PR author.

The error messages are thus informational only, and we will figure out a
plan to fail if the actual delta before and after a PR is large.