Turn off service account for istioctl test (#1474)
I am pretty sure this is not needed as the test runs fully locally now
with kind, so it doesnt need boskos or gcr. The istioctl test is already
failing (https://github.com/istio/istio/issues/15500) so this will
reduce risk if this doesn't work as expected. Once this works others can
be removed as well