Raise job CPU limits
The CPU limits were set to a low number in an attempt to fix
infrastructure flakes we attributed to CPU being overloaded. When we
made this change, we did not detect any noticable differences, and it
appears moving to SSDs has fixed the issue. Right now our tests get
heavily throttled. Hopefully this should help.