Finish modernization of operator
As per @fejta's PR: https://github.com/istio/test-infra/pull/1502

This moves the rest of the jobs to the model outlined by @fejta.  Please
use this structure when adding new jobs for the operator repo.  When Kind
becomes a necessity, please consult either @fejta or @sdake and we will
get you moving (or you can sort it out on your own if you prefer :)