Delete istio-pilot-e2e-envoyv2-v1alpha3-k8s-latest (#1630)
This test is completely irrelevant now. It seems the intent was to
rerun an existing test with the "latest kubernetes", with the latest
meaning 1.11. The normal test now actually runs on kubernetes latest
(1.15), and we have explicit tests to test all supported versions.
Therefor, it seems this test provides no additional coverage.