Set BUILD_WITH_CONTAINER to 0 for all jobs. (#1676)
By defaults, we want humans building Istio stuff to be building through a
container. But the CI system already cranks up jobs within said container,
so we never want to have the CI's call to 'make' to induce a dind call
to the container. So we set BUILD_WITH_CONTAINER to 0 for all CI jobs.