Temporarily disable CNI test
The CNI tests were accidentally broken in a manner that is likely not
fixable in a day (partly because they depend on nightly builds which are
only built once a day - this is another problem..., partly because I
don't know the best fix). We just made this test required, so I don't
think its too harmful to make it optional again and unblock PRs.

Tracking in https://github.com/istio/istio/issues/16508