Add postsubmit test for Kubernetes 1.16
Attempting to be a bit proactive with our testing here and start testing
this before our customers do. Unfortunately kind doesn't publish
prerelease images (or I couldn't find them), but it was pretty trivial
to build one myself for 1.16 beta 1.

I did manually test this on a few tests and things seem to work.