Kill test-infra-update-deps
This script has not succeeded for as far back as prow history goes, and
no one has missed it, so it seems it was not actually providing value.

If it ever was useful, I doubt it is relevant now as our codebase has
changed enough that it would not be compatible (go modules). If we ever
do want this again in the future we can always bring it back, but
I suspect there are much better ways to automatically update
dependencies.

Fixes https://github.com/istio/istio/issues/15965