Update testgrid for job name changes (#1781)
* Update testgrid for job name changes

We update a bunch of job names making the config out dated. In the
process, we added a bunch of jobs to the testgrid. This PR regenerates
the config for these jobs, and moves the formatting around a bit so
there is a tab for each repo rather than each repo-branch, since we
would have a lot of tabs otherwise.

I had to explicitly disable the two trusted test-infra jobs for now --
not sure why I couldn't get them to work. In a followup I can try to fix
them, but this PR is already large enough.

* Fix test