Refactor testgrid layout
Currently, our testgrid has so many tests in it that it is hard to see -
the tabs take up half of the screen, and it will only get worse. This
changes the names to used _ as a delimiter, like test names, for
clarity, and splits the different release branches to different tabs.
Additionally, it makes postsubmits come first, as these are the things
developers typically care about -- I have heard multiple people get
confused by looking at presubmits accidentally.