Bump test CPU requests
We have been seeing an increasing number of infrastructure flakes. It
seems likely this is related to overloaded pods as it is correlated to
times when large number of tests are running. Bumping the requests up
should give us some more consistent behavior.

Note that as we scaled down boskos the prow cluster load has increased.