Improve job specs
- Remove support for suppress_entrypoint. Instead, explicitly specify the entrypoint
command when needed. This is a bit less magic, and simplifies most cases.

- Make the default image be the build-tools image rather than istio-builder, which
means upgrading to a new image requires making one central change.

- Support not specifying any branches in a job config. That defaults it to master.

- Introduce the `branch` command-line option. This will automatically create the job
configs for a release branch. It'll create new job config files which snapshot the
build image. This is paired with the new supports_release_branching field in the job
config, to indicate which sets of jobs to clone for a release. You just invoke the
generator with `go run generate.go branch 1.4" in order to get the 1.4 set of jobs.