Move proxy to prow config gen
This PR moves the proxy to the new prow config generation. I made this
enabled only for the "prow-stage" branch so we can do send a dummy PR to
validate I didn't break anything.

Benefits:

Proxy follows same conventions of other repos (in terms of naming,
testgrid, branching, etc)
Config should be much simpler to understand
Resolves some issues with duplicate names, confusing context, etc
The only effective changes after this is the change to the job names. I
am open to renaming, the reason I didn't keep the job names as before is
we would have ended up with names like
proxy-postsubmit_proxy_postsubmit, due to the naming convention of
test_repo_jobtype.