Run dry run publish only if code changes
Right now we run a ~30min test of a dry run build as part of presubmit.
This makes building and publishing official releases painful.

Instead, only run when the actual release logic changes.

Additionally, rename this to dry-run. `publish` is really confusing name
when we also have a real publish.