Fix release branching image
Previously we had an outdated, hardcoded image as a fallback in the
generation code. Every job defines the image they use, so we don't need
a fallback. The "branch" command was using this image instead of what it
was branching from.

This removes the fallback, which in turn fixes the 1.5 jobs to use the
latest image (they fail right now since they use old images).