Stop caching go build results (#2571)
This is the go build cache, which caches tests and builds. I would be
fine with having builds cached but I am pretty hesitant to be having our
tests be cached and worry that it won't properly be testing changes as
the cache is not always properly invalidated by yaml changes, etc. It
also leads to erratic test durations, which is hard to track trends. It
also is not very effective, as all the concurrent jobs invalidate the
cahce often. I also couldn't find anyone else doing this.