Another attempt at go module caching (#2813)
previously we tried NFS volumes:
https://github.com/istio/test-infra/issues/2731

This has two issues, one that NFS is slow and 2 that there was
apparently a race at pod startup.

This PR once again tries to add caching, but this time using HostPaths.
I add 2 hidden/skipped jobs to experiment with this. In addition, I
removed the 1.6 jobs using the NFS mount. This will ensure they are not
impacted by this change, and will allow me to delete the NFS drive (its
kind of expensive)