Increase scheduling timeout (#2818)
Unlike K8s prow (I think), we have autoscaling on our build cluster. As
a resultm during spikes in tests we see pod scheduling timeouts, as the
cluster autoscaling can take a few minutes for large increases. As a
result the tests fail. I don't see a reason to fail fast when waiting is
very likely to be succesful, so I propose we wait longer.