Authentikos: Refactor Token/Secret creators to interfaces.
This allows for other token types to be created while reusing the same
common creation / reconcilation components.

As an example for how this can be extended, adds a GitHub App token
creator as an alternative to the existing Google OAuth implementation.

Also changes secret updates to patches of just the data to fix an issue
dicovered during testing where the reconciler would overwrite any
user-added labels/annotations.