Authentikos: Fix validation test when running as root.
If running as root (i.e. default user for docker run), setting the file
permissions to `0000` doesn't actually prevent the file from being
unreadable. This was causing false signals in CI testing. This modifies
the behavior to just create a file that doesn't exist, which has the
same effect of what we're trying to test (e.g. that we can't read the
template file).

Since the deletedCredsFile is similar in this sense this was also
changed to be consistent.

Also added some minor test QoL changes (usage of t.Helper,
t.Log) to aid future debugging.