testgrid: dashboards must belong to dashboard groups (#4356)
* testgrid: dashboards must belong to dashboard groups

The goal here is to prevent istio from littering testgrid.k8s.io with a
bunch of top-level dashboards. We follow a convention there of placing
all dashboards under a group.

This adds a minimal version of the testgrid config test used in
kubernetes/test-infra. It verifies that testgrid config parses
correctly, and then verifies the following conventions:

- dashboard and dashboard group names must start with "istio"
- dashboards must belong to a dashboard group
- dashboards should be in the group with the longest common prefix

With the test in place, the config failed, so the config was then
modified to pass the test:

- add istio_release-1.16 group
- move other dashboards to appropriate groups

* testgrid: lint fixes

Specifically:
- gci write testgrid/config_test.go
- gofumpt -w testgrid/config_test.go

* licenses: add bitbucket.org/creachadair/stringset