Rotate test-infra-trusted build cluster kubeconfig (#4373)
The test-infra-trusted cluster root CA is expiring soon, which will need to be rotated, and the old kubeconfig for Prow service cluster to authenticate with this build cluster will be invalid after the rotation. So a new kubeconfig is needed to avoid breakage of the connection to this build cluster.

Due to security reasons GKE doesn't allow long lived tokens any more, and the newly generated tokens are at most valid for 2 days, so there needs a way to rotate the token. This PR follows the strategy adopted by k8s prow