prow: migrate from gencred to gke-gcloud-auth-plugin/WI
Currently, we have a job the runs every so often and updates our
kubeconfig Secrets with a new token. This is not ideal from a security
and reliability POV.

This PR migrates off of this onto `gke-gcloud-auth-plugin` which uses
workload identity in the prow control plane components to request access
tokens on-demand.

This PR:
* Adds the new kubeconfig configmap. It has no private info so its no
  longer a Secret
* Removes the old ExternalSecrets, gencred job, gencred config
* Updates the prow control plane to reference the new configs

I did a dry run of this by updating one of the clusters to this method
to validate permissions are setup correctly, etc.