Remove cosign 'secret' (#4788)
This is not a secret, its just a simple value.

To keep things simple, we remove the secret and inline it.

This is part of an effort to remove all secrets from the cluster. While
this one technically is fine, I think it causes confusion that it is
confidential when it isn't.