Create generate-deps-configmaps script
The purpose of this script is to generate the istio-deps and
release-deps configmaps that are needed for istio-private work.

Currently these are not generated and are captured after-the-fact when
the PSWG team is building security releases. We should, ideally, have
this as part of the branch cutting process. Cleanup will still be
manual, but can occur during the regular EOL cleanup flow.

Signed-off-by: Daniel Hawton <daniel.hawton@solo.io>