Create generate-deps-configmaps tool (#5186)
* Create generate-deps-configmaps script

The purpose of this script is to generate the istio-deps and
release-deps configmaps that are needed for istio-private work.

Currently these are not generated and are captured after-the-fact when
the PSWG team is building security releases. We should, ideally, have
this as part of the branch cutting process. Cleanup will still be
manual, but can occur during the regular EOL cleanup flow.

Signed-off-by: Daniel Hawton <daniel.hawton@solo.io>

* Update tools/generate-deps-configmaps/README.md

Co-authored-by: Ian Rudie <ilrudie@gmail.com>

* Update tools/generate-deps-configmaps/README.md

* Update tools/generate-deps-configmaps/generate.py

Co-authored-by: Ian Rudie <ilrudie@gmail.com>

* Update tools/generate-deps-configmaps/generate.py

Co-authored-by: Ian Rudie <ilrudie@gmail.com>

* Update tools/generate-deps-configmaps/generate.py

Co-authored-by: Ian Rudie <ilrudie@gmail.com>

* error handling, readd comments, and fix branch names in existing istio-deps yaml

---------

Signed-off-by: Daniel Hawton <daniel.hawton@solo.io>
Co-authored-by: Ian Rudie <ilrudie@gmail.com>