prow: protect all branches in repos
This changes our configuration to protect *all* branches in repos, not
just release ones.

This means users will need to work in forks. When operating from the UI,
GitHub should automatically do this for them. For development with
standard CLI tooling, users will need to adopt their workflows to
accomodate.

The `branches` config should not be required anymore, but I left it just
in case this doesn't work how I expect it to. Once we validate it, we
can remove those as well.